Category repo
Former-commit-id: d15b761a510c1bf0d4e6c2593151b527763d6b39 [formerly 47ad3a6101c439cd183318c6cbf599af6d3b89ae]
Former-commit-id: 1bb24f2b767ba0d6a63840e3310c50ca2f3e3c87

+++ api/domain/categories.go
@@ -13,7 +13,7 @@ type (
 	// Category defines the groups used for categorising
 	// individual posts.
 	Category struct {
-		Id          int       `db:"id" json:"id"`
+		Id          int64     `db:"id" json:"id"`
 		UUID        uuid.UUID `db:"uuid" json:"uuid"`
 		Slug        string    `db:"slug" json:"slug" binding:"required,max=150"`
 		Name        string    `db:"name" json:"name" binding:"required,max=150"`
